fix(reports): report the numbers the shop can check
Five figures disagreed with the data behind them.

«Доля в закупках» divided each product by the sum of the rows that fit on
the page, not by the period's spend — with the top 50 covering 13589.83 of
19406.22, every share read too big.

A voided purchase line leaves its receipt standing and reverses only the
layer it wrote. PO #56 kept 28 of those, and the purchase report billed the
shop for goods the ledger had already sent back.

Top products carried quantity_sold as an int, so 43.142 kg of potatoes sold
as 43, and their profit came from today's cost card rather than the cost at
the moment of sale — the same goods showed one profit here and another in
the profit report. The reports page then printed total_profit, a field the
API does not send, so it always showed 0.

«30 дней» ran from today minus 30 to today, which is 31 days, while the MCP
connector's last_30_days is 30. Both mean 30 now.

+++ sellary-backend/schemas/report.py
@@ -44,7 +44,7 @@ class TopProductItem(BaseModel):
     product_id: int
     product_name: str
     barcode: Optional[str] = None
-    quantity_sold: int
+    quantity_sold: Decimal
     revenue: Decimal
     profit: Decimal
